DELETE /entries/{serialNumber} is restricted to admin

+++ src/main/kotlin/com/github/hmiyado/kottage/route/entries/EntriesSerialNumberLocation.kt
@@ -1,18 +1,14 @@
 package com.github.hmiyado.kottage.route.entries
 
-import com.github.hmiyado.kottage.authentication.UserPrincipal
 import com.github.hmiyado.kottage.openapi.apis.OpenApi
 import com.github.hmiyado.kottage.route.Path
 import com.github.hmiyado.kottage.route.allowMethods
 import com.github.hmiyado.kottage.service.entries.EntriesService
 import io.ktor.application.call
-import io.ktor.auth.authenticate
-import io.ktor.auth.authentication
 import io.ktor.http.HttpMethod
 import io.ktor.http.HttpStatusCode
 import io.ktor.locations.KtorExperimentalLocationsAPI
 import io.ktor.locations.Location
-import io.ktor.locations.delete
 import io.ktor.locations.get
 import io.ktor.locations.options
 import io.ktor.response.respond
@@ -57,15 +53,12 @@ data class EntriesSerialNumberLocation(val serialNumber: Long) {
                             }
                         }
                 }
-            }
-
-            authenticate("user") {
-                delete<EntriesSerialNumberLocation> { (serialNumber) ->
-                    val (user) = call.authentication.principal<UserPrincipal>() ?: kotlin.run {
-                        call.respond(HttpStatusCode.Unauthorized)
-                        return@delete
+                entriesSerialNumberDelete { user ->
+                    val serialNumber = call.parameters["serialNumber"]?.toLongOrNull()
+                    if (serialNumber == null) {
+                        call.respond(HttpStatusCode.BadRequest)
+                        return@entriesSerialNumberDelete
                     }
-
                     kotlin.runCatching { entriesService.deleteEntry(serialNumber, user.id) }
                         .onSuccess {
                             call.respond(HttpStatusCode.OK)
